Remove StorageCollection.currentIndex
The concept behind the current index was very strange and unpredictable.

Why it was unpredictable?

I can create a MutableStorage with an index equal to StorageCollection.currentIndex and it points to a particular offset in the savedata that is the CurrentStorageOffset. If I change che currentIndex value, my MutableStorage instance will still has the old index but it will points to a different data chunk. So I should invalidate the old Storage instance but this must be done in the UI layer and it's error prone